move all time utils to same file, remove unused stuff
