+++ src/components/CourseForm/CourseForm.tsx
@@ -26,7 +26,7 @@ import { useMessage } from '../Providers/MessageProvider';
 import { post, update } from '@/lib/response/fetchUtil';
 import FormFieldError from '../FormFieldError/FormFieldError';
 import { Tag } from '@prisma/client';
-import { dateToDateTimeLocal } from '@/lib/util';
+import { dateToDateTimeLocal } from '@/lib/timedateutils';
 import { CourseWithTags } from '@/lib/prisma/courses';
 
 type FormType = CourseSchemaType | CourseSchemaWithIdType;
